Refactored batchSize to only default in one place
